src/main/java/creational/sigleton/DoubleCheckLockingSingleton.java
package creational.sigleton;

public class DoubleCheckLockingSingleton {
    private static volatile DoubleCheckLockingSingleton instance;

    private DoubleCheckLockingSingleton() {
        // protect against reflection
        if (instance != null) {
            throw new RuntimeException("Use getInstance to instantiate this class.");
        }
    }

    public static DoubleCheckLockingSingleton getInstance() {
        if (instance == null) {
            synchronized (DoubleCheckLockingSingleton.class) {
                if (instance == null) {
                    instance = new DoubleCheckLockingSingleton();
                }
            }
        }
        return instance;
    }
}
